simpleCrud/simple-crud-api-data/src/main/java/com/portofolio/demo/models/json/stock/Stock.java
package com.portofolio.demo.models.json.stock;

public class Stock {
    private Long id;
    private int quantity;
    private String itemName;
    private String uri;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public int getQuantity() {
        return quantity;
    }

    public void setQuantity(int quantity) {
        this.quantity = quantity;
    }

    public String getItemName() {
        return itemName;
    }

    public void setItemName(String itemName) {
        this.itemName = itemName;
    }

    public String getUri() {
        return uri;
    }

    public void setUri(String uri) {
        this.uri = uri;
    }
}
